Adding GraphViz Dependency
Adding a GraphViz dependency to the NIEM documentation site. We'd like to create GraphViz graphics inline in the site code without rendering in a seperate program. There is a plugin specified in the Gemfile for the Jekyll/GraphViz interface using the GraphViz-dot program.

Made  a small modification to the GitHub pages README file to directly specify that this project depends on Ruby and GraphViz being installed on the hosting machine; as well as fixing the commands to run the site from your local machine through bundler.